Fixes lp:6080191. Passes the --distribution switch to python-mkdebian if specified in quickly configure.
Points to note: I have used <ubuntu-distribution-name> in the help message when
there are a wrong number of arguments for quickly configure
target_distribution. This is because there is no support Debian on Launchpad
PPAs as yet.

Also, the --distribution switch was merged into python-distutils-extra's trunk
today. So you may want to wait before merging this.